backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-socketio
python-dotenv
psycopg2-binary
